+++ packages/s2-core/src/interaction/data-cell-brush-selection.ts
@@ -7,6 +7,7 @@ import {
   InteractionStateName,
 } from '../common/constant/interaction';
 import type { BrushRange, ViewMeta } from '../common/interface';
+import { updateRowColCells } from '../utils';
 import { BaseBrushSelection } from './base-brush-selection';
 
 /**
@@ -71,16 +72,19 @@ export class DataCellBrushSelection extends BaseBrushSelection {
         colIndex < range.end.colIndex + 1;
         colIndex++
       ) {
-        const colId = colLeafNodes[colIndex].id;
+        const colId = String(colLeafNodes[colIndex].id);
         let rowId = String(rowIndex);
         if (rowLeafNodes.length) {
-          rowId = rowLeafNodes[rowIndex].id;
+          rowId = String(rowLeafNodes[rowIndex].id);
         }
         metas.push({
           colIndex,
           rowIndex,
           id: `${rowId}-${colId}`,
           type: 'dataCell',
+          rowId,
+          colId,
+          spreadsheet: this.spreadsheet,
         });
       }
     }
@@ -89,14 +93,22 @@ export class DataCellBrushSelection extends BaseBrushSelection {
 
   // 最终刷选的cell
   protected updateSelectedCells() {
-    const { interaction } = this.spreadsheet;
+    const { interaction, options } = this.spreadsheet;
 
     const range = this.getBrushRange();
+    const selectedCellMetas = this.getSelectedCellMetas(range);
 
     interaction.changeState({
-      cells: this.getSelectedCellMetas(range),
+      cells: selectedCellMetas,
       stateName: InteractionStateName.SELECTED,
     });
+
+    if (options.interaction.selectedCellHighlight) {
+      selectedCellMetas.forEach((meta) => {
+        updateRowColCells(meta);
+      });
+    }
+
     this.spreadsheet.emit(
       S2Event.DATA_CELL_BRUSH_SELECTION,
       this.brushRangeCells,
